initialize _displacementSite to NULL

+++ compiler/compile/DisplacementSites.cpp
@@ -1,7 +1,7 @@
 #include "compile/Compilation.hpp"
 #include "compile/DisplacementSites.hpp"
 
-TR_DisplacementSite::TR_DisplacementSite(TR::Compilation *comp, uint32_t assumptionID)
+TR_DisplacementSite::TR_DisplacementSite(TR::Compilation *comp, uint64_t assumptionID)
       : _assumptionID(assumptionID), _calleeIndex(0),
 	_byteCodeIndex(0), _dispSize(TR_DisplacementSite::bits_32), _location(NULL)
    {
